Initial commit
This repository is the builder etc for blt/bughunt-core, an
implementation of the bare skeleton of what it takes to get LLVM trace
information flowing into Rust. The area of most importance is
src/lib.rs where the rustc flags live.

No attempt has been made to compile this on anything other than OS X.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>